third_party/blink/web_tests/platform/generic/svg/dom/SVGStringList-expected.txt

This test checks the SVGStringList API - utilizing the requiredExtensions property of SVGRectElement

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Check initial requiredExtensions values
PASS rect.requiredExtensions.numberOfItems is 2
PASS rect.requiredExtensions.getItem(0) is "foo"
PASS rect.requiredExtensions.getItem(1) is "bar"


Check that getItem() does NOT return live strings, as the IDL defines the return types as plain DOMString
PASS firstItem = 'test' is "test"
PASS rect.requiredExtensions.numberOfItems is 2
PASS rect.requiredExtensions.getItem(0) is "foo"
PASS rect.requiredExtensions.getItem(1) is "bar"
PASS successfullyParsed is true

TEST COMPLETE
